Update filtermap benchmarks
Benchee won't give accurate results if you don't wrap your code in a
precompiled module. Or else the interpretation time of the lambda is
taken into account, which makes things like macros seem much slower.

Also, making the input "random" can help, as the compiler won't be able
to optimize for it.